+++ src/features/capsule-answered/components/capsule-share-template.tsx
@@ -1,5 +1,6 @@
 import { Text } from '@/src/components/texts';
 import { Colors } from '@/src/constants/theme';
+import { Corner } from '@/src/features/capsule-answered/components/corner';
 import { CapsuleResponse } from '@/src/features/capsule-reponse/types/capsule-response.types';
 import { Image, View } from 'react-native';
 
@@ -13,67 +14,66 @@ export const CapsuleShareTemplate = ({ capsuleResponse }: CapsuleShareTemplatePr
             style={{
                 width: 1080,
                 height: 1920,
-                backgroundColor: Colors.linenCloud,
-                paddingHorizontal: 140,
-                paddingVertical: 180,
+                backgroundColor: Colors.slateRoot,
+                padding: 90,
                 justifyContent: 'space-between',
-                borderWidth: 1,
+                gap: 32
             }}
         >
-            <View
-                style={{
-                    alignItems: 'center',
-                    gap: 40,
-                }}
-            >
+            <View style={{ alignItems: 'center', gap: 40 }}>
                 <Image
-                    source={require('@/assets/images/selen-without-background.png')}
-                    style={{
-                        width: 250,
-                        height: 250,
-                        borderRadius: 150,
-                    }}
+                    source={require('@/assets/images/icon_selen_512.png')}
+                    style={{ width: 300, height: 300, borderRadius: 150 }}
                 />
-                <Text
-                    style={{
-                        fontSize: 48,
-                        color: Colors.oakHoney,
-                        letterSpacing: 2,
-                    }}
-                >
+                <Text style={{ fontSize: 48, color: Colors.linenCloud, letterSpacing: 2 }}>
                     CAPSULE DU JOUR
                 </Text>
             </View>
 
-            <View style={{ gap: 80 }}>
-                <Text
-                    style={{
-                        fontSize: 56,
-                        color: Colors.slateRoot,
-                        textAlign: 'center',
-                        lineHeight: 84,
-                    }}
-                >
-                    {capsuleResponse?.aiResponse}
-                </Text>
+            <View
+                style={{
+                    position: 'relative',
+                    backgroundColor: Colors.linenCloud,
+                    borderRadius: 32,
+                    paddingHorizontal: 80,
+                    paddingVertical: 80,
+                }}
+            >
+                <Corner style={{ top: 24, left: 24, transform: [{ rotate: '270deg' }] }} color={Colors.slateRoot} />
+                <Corner style={{ top: 24, right: 24, transform: [{ rotate: '0deg' }] }} color={Colors.slateRoot} />
+                <Corner style={{ bottom: 24, right: 24, transform: [{ rotate: '90deg' }] }} color={Colors.slateRoot} />
+                <Corner style={{ bottom: 24, left: 24, transform: [{ rotate: '180deg' }] }} color={Colors.slateRoot} />
+
+                <View style={{ gap: 40 }}>
+                    <View style={{ flexDirection: 'row', alignItems: 'center', justifyContent: 'center', gap: 16 }}>
+                        <View style={{
+                            backgroundColor: Colors.sateRootLight,
+                            borderRadius: 100,
+                            width: 64,
+                            height: 64,
+                            alignItems: 'center',
+                            justifyContent: 'center',
+                        }}>
+                            <Text style={{ fontSize: 28 }}>🌙</Text>
+                        </View>
+                        <Text variant="bold" family="seasons" style={{ fontSize: 52, color: Colors.slateRoot }}>
+                            La réponse de la lune
+                        </Text>
+                    </View>
+                    <Text style={{ fontSize: 40, color: Colors.slateRoot, textAlign: 'center' }}>
+                        &quot;{capsuleResponse?.aiResponse}&quot;
+                    </Text>
+                </View>
             </View>
 
-
             <View style={{ gap: 8 }}>
-                <Text
-                    style={{
-                        fontSize: 40,
-                        color: Colors.oakHoneyDark,
-                        textAlign: 'center',
-                    }}
-                >
+                <Text style={{ fontSize: 40, color: Colors.linenCloud, textAlign: 'center' }}>
                     Prêt(e) à dompter ton chaos ? Découvre ta capsule du jour sur l&apos;app Selen
                 </Text>
-                <Text style={{ fontSize: 32, textAlign: 'center', }}>
+                <Text style={{ fontSize: 32, textAlign: 'center', color: Colors.linenCloud }}>
                     Lien dans la bio @selen_app_officiel
                 </Text>
             </View>
-
         </View>
     );
 };